feat: 진도율 추가

+++ src/main/java/BookMind/GraduationProject_BE/DTO/UserBookDTO.java
@@ -18,7 +18,7 @@ public class UserBookDTO {
     // UserBook과 관련된 필드들
     private String status; // 읽기 상태 (READING, COMPLETED)
     private Boolean favorite; // 즐겨찾기 여부
-    private int lastReadPage; // 마지막으로 읽은 페이지
+    private float lastReadPage; // 마지막으로 읽은 페이지 (진도율)
     private String startDate; // 시작일
     private String endDate; // 완료일
     private Byte rating; // 평점
